Refactor query helpers into top-level module
